Set alarms to fire at actual times
Updated the next fire time for all alarms when saving.
Determined which alarm fires next if any.
Set a real time alert for the next alarms time.
Added a boot receiver so the alarm sounds even after rebooting.
Added a check so the alert activity doesn't immediately pause if
launched while the screen is locked.